DISPATCH-161 - Move all annotation related functions into annotation.h with corresponding implementation in annotation.c
