+++ include/qpid/dispatch/message.h
@@ -132,66 +132,6 @@ void qd_message_free(qd_message_t *msg);
 qd_message_t *qd_message_copy(qd_message_t *msg);
 
 /**
- * Retrieve the message annotations from a message.
- *
- * IMPORTANT: The pointer returned by this function remains owned by the message.
- *            The caller MUST NOT free the parsed field.
- *
- * @param msg Pointer to a received message.
- * @return Pointer to the parsed field for the message annotations.  If the message doesn't
- *         have message annotations, the return value shall be NULL.
- */
-qd_parsed_field_t *qd_message_message_annotations(qd_message_t *msg);
-
-/**
- * Set the value for the QD_MA_TRACE field in the outgoing message annotations
- * for the message.
- *
- * IMPORTANT: This method takes ownership of the trace_field - the calling
- * method must not reference it after this call.
- *
- * @param msg Pointer to an outgoing message.
- * @param trace_field Pointer to a composed field representing the list that
- * will be used as the value for the QD_MA_TRACE map entry.  If null, the
- * message will not have a QA_MA_TRACE message annotation field.  Ownership of
- * this field is transferred to the message.
- *
- */
-void qd_message_set_trace_annotation(qd_message_t *msg, qd_composed_field_t *trace_field);
-
-/**
- * Set the value for the QD_MA_TO field in the outgoing message annotations for
- * the message.
- *
- * IMPORTANT: This method takes ownership of the to_field - the calling
- * method must not reference it after this call.
- *
- * @param msg Pointer to an outgoing message.
- * @param to_field Pointer to a composed field representing the to overrid
- * address that will be used as the value for the QD_MA_TO map entry.  If null,
- * the message will not have a QA_MA_TO message annotation field.  Ownership of
- * this field is transferred to the message.
- *
- */
-void qd_message_set_to_override_annotation(qd_message_t *msg, qd_composed_field_t *to_field);
-
-/**
- * Set the value for the QD_MA_INGRESS field in the outgoing message
- * annotations for the message.
- *
- * IMPORTANT: This method takes ownership of the ingress_field - the calling
- * method must not reference it after this call.
- *
- * @param msg Pointer to an outgoing message.
- * @param ingress_field Pointer to a composed field representing ingress router
- * that will be used as the value for the QD_MA_INGRESS map entry.  If null,
- * the message will not have a QA_MA_INGRESS message annotation field.
- * Ownership of this field is transferred to the message.
- *
- */
-void qd_message_set_ingress_annotation(qd_message_t *msg, qd_composed_field_t *ingress_field);
-
-/**
  * Receive message data via a delivery.  This function may be called more than once on the same
  * delivery if the message spans multiple frames.  Once a complete message has been received, this
  * function shall return the message.
